lambda-functions/agent-lambda/requirements.txt
boto3
python-dotenv
supabase==2.24.0